Make the buffalo verification use the "help" subcommand
`buffalo` with no subcommand returns an error after printing out the help message

It's a teensy thing, but I think it's a better look not to show new users an error message when everything is really ok.

I also updated the output of the help command to the latest version.

I swear to frick that I'm not trying to "score points". I just think these small introductory things are really good to get right.